requirements_vector.txt
openai>=1.0.0
pymongo>=4.6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
numpy>=1.24.0

